10/31

+++ src/Estimation.java
@@ -5,8 +5,8 @@ public abstract  class Estimation {
 
     public abstract void EstimateSchedule(Schedule s);
 
-    public abstract double Estimate_Schedule_Time(Schedule s);
+    public abstract float Estimate_Schedule_Time(Schedule s);
 
-    public abstract double Estimate_Schedule_Money(Schedule s);
+    public abstract float Estimate_Schedule_Money(Schedule s);
 
 }
